feat: add routing and layout scaffolding
