+++ package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "tsc && vite build",
+    "build": "vite build",
     "preview": "vite preview",
     "tauri": "tauri"
   },
@@ -57,6 +57,7 @@
     "react": "^18.3.1",
     "react-day-picker": "^9.8.0",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.22.3",
     "react-hook-form": "^7.60.0",
     "react-resizable-panels": "^3.0.3",
     "recharts": "^3.0.2",
