+++ src/pages/AnalyticsPage.tsx
@@ -1,0 +1,8 @@
+const AnalyticsPage = () => (
+  <div className="p-6">
+    <h1 className="text-2xl font-bold mb-4">Analytics</h1>
+    <p>Analytics dashboard coming soon...</p>
+  </div>
+);
+
+export default AnalyticsPage;
