+++ src/pages/KanbanPage.tsx
@@ -1,0 +1,8 @@
+const KanbanPage = () => (
+  <div className="p-6">
+    <h1 className="text-2xl font-bold mb-4">Kanban</h1>
+    <p>Kanban board coming soon...</p>
+  </div>
+);
+
+export default KanbanPage;
