+++ src/pages/TimelinePage.tsx
@@ -1,0 +1,8 @@
+const TimelinePage = () => (
+  <div className="p-6">
+    <h1 className="text-2xl font-bold mb-4">Timeline</h1>
+    <p>Timeline view coming soon...</p>
+  </div>
+);
+
+export default TimelinePage;
